appfiles-main/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask
flask-cors
Flask-Mail
Flask-Bcrypt
itsdangerous
python-dotenv
psycopg2-binary
